+++ src/main/CsharpDotNet2/LaunchDarkly/Api/Api/UsersApi.cs
@@ -29,7 +29,7 @@ namespace LaunchDarkly.Api.Api
         /// <param name="offset">Specifies the first item to return in the collection.</param>
         /// <param name="after">A timestamp filter, expressed as a Unix epoch time in milliseconds. All entries returned will have occured after this timestamp.</param>
         /// <returns>Users</returns>
-        Users GetSearchUsers (string projectKey, string environmentKey, string q, int? limit, int? offset, decimal? after);
+        Users GetSearchUsers (string projectKey, string environmentKey, string q, int? limit, int? offset, long? after);
         /// <summary>
         /// Get a user by key. 
         /// </summary>
@@ -160,7 +160,7 @@ path = path.Replace("{" + "userKey" + "}", ApiClient.ParameterToString(userKey))
         /// <param name="offset">Specifies the first item to return in the collection.</param> 
         /// <param name="after">A timestamp filter, expressed as a Unix epoch time in milliseconds. All entries returned will have occured after this timestamp.</param> 
         /// <returns>Users</returns>            
-        public Users GetSearchUsers (string projectKey, string environmentKey, string q, int? limit, int? offset, decimal? after)
+        public Users GetSearchUsers (string projectKey, string environmentKey, string q, int? limit, int? offset, long? after)
         {
             
             // verify the required parameter 'projectKey' is set
